revert filter ui, usefiltermodal as prop
