+++ components/search-filter-modal.tsx
@@ -2,11 +2,9 @@
 /* eslint-disable @typescript-eslint/no-unsafe-call */
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
-import { ChangeEventHandler, Fragment, useMemo, useRef } from "react";
+import { ChangeEventHandler, Fragment, useRef } from "react";
 
-import { FormLabel } from "./ui/forms/form-label";
-import { InputSelect } from "./ui/forms/input-select";
-import { SelectSkeleton } from "./ui/skeleton-loading";
+import { SearchFilter } from "~/components/search-filter";
 
 import { Dialog, Transition } from "@headlessui/react";
 import { XIcon } from "@heroicons/react/solid";
@@ -17,10 +15,10 @@ export interface SortSetting {
 }
 
 export interface SearchFilterModalProps {
-  isLoading?: boolean;
   checkDocSize?: boolean;
   filters: any;
   filterItems?: {};
+  isLoading?: boolean;
   isOpen: boolean;
   handleFilterChange: ChangeEventHandler<HTMLSelectElement>;
   handleSortChange: ChangeEventHandler<HTMLSelectElement>;
@@ -42,110 +40,6 @@ export function SearchFilterModal({
   sortSettings,
 }: SearchFilterModalProps) {
   const closeButtonRef = useRef(null);
-
-  const filterForms = useMemo(
-    () =>
-      filterItems && (
-        <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
-          {Object.entries(filterItems).map(([key, value], idx) => {
-            const { title, buckets }: any = value;
-            return (
-              <div key={`filter-${idx}`} className="space-y-1">
-                <FormLabel htmlFor={`filter-${key}`}>{title}</FormLabel>
-                <InputSelect
-                  name={key}
-                  onChange={handleFilterChange}
-                  title={title}
-                  value={filters?.[key]?.length ? filters[key][0] : ""}
-                >
-                  <option value="">Semua</option>
-                  {buckets.map((bucket: any, bIdx: number) => {
-                    if (bucket.key) {
-                      if (checkDocSize) {
-                        if (bucket.doc_count > 0) {
-                          return (
-                            <option
-                              key={`option-${key}-${bIdx + 1}`}
-                              value={bucket.key}
-                            >
-                              {bucket.key}
-                            </option>
-                          );
-                        }
-                        // Do not print, when doc_count = 0 and checkDocSize set to true
-                        return null;
-                      } else {
-                        // FAQ page doesn't check the doc_count
-                        // just pass checkDocSize props to false
-                        return (
-                          <option
-                            key={`option-${key}-${bIdx + 1}`}
-                            value={bucket.key}
-                          >
-                            {bucket.key}
-                          </option>
-                        );
-                      }
-                    }
-                    return null;
-                  })}
-                </InputSelect>
-              </div>
-            );
-          })}
-        </div>
-      ),
-    [filterItems, handleFilterChange, filters, checkDocSize],
-  );
-
-  const sortForms = useMemo(
-    () =>
-      sortSettings?.length && (
-        <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
-          <div className="space-y-1">
-            <FormLabel htmlFor="sort-by">Urut berdasarkan</FormLabel>
-            <InputSelect
-              name="sort-by"
-              onChange={handleSortChange}
-              title="Urut berdasarkan"
-              value={sortBy}
-            >
-              {sortSettings.map((cur, idx) => {
-                return (
-                  <option key={`sort-by-${idx}`} value={cur.value}>
-                    {cur.label}
-                  </option>
-                );
-              })}
-            </InputSelect>
-          </div>
-        </div>
-      ),
-    [handleSortChange, sortBy, sortSettings],
-  );
-
-  const renderFilterForms = () => {
-    if (isLoading) {
-      return (
-        <>
-          <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
-            <SelectSkeleton />
-          </div>
-          <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
-            <SelectSkeleton />
-            <SelectSkeleton />
-          </div>
-        </>
-      );
-    }
-
-    return (
-      <>
-        {filterForms}
-        {sortForms}
-      </>
-    );
-  };
 
   return (
     <Transition.Root as={Fragment} show={isOpen}>
@@ -206,7 +100,18 @@ export function SearchFilterModal({
                   >
                     Filter
                   </Dialog.Title>
-                  <div className="mt-4 space-y-4">{renderFilterForms()}</div>
+                  <div className="mt-4 space-y-4">
+                    <SearchFilter
+                      checkDocSize={checkDocSize}
+                      filterItems={filterItems}
+                      filters={filters}
+                      handleFilterChange={handleFilterChange}
+                      handleSortChange={handleSortChange}
+                      isLoading={isLoading}
+                      sortBy={sortBy}
+                      sortSettings={sortSettings}
+                    />
+                  </div>
                 </div>
               </div>
             </div>
